Update tests + add isDotCom prop
